Fixed error redefinition of typedef 'FSCalendarMonthPosition'
